refactor: Renamed gqlschema to graphqlgo and renamed go package from gqlgraphqlator to graphqlgo
