+++ substancegen/generators/graphqlgo/graphqlgo.go
@@ -1,4 +1,4 @@
-package gqlgraphqlator
+package graphqlgo
 
 import (
 	"fmt"
