+++ substancegen/generators/graphqlgo/graphqlgoTemplates.go
@@ -1,0 +1,14 @@
+package graphqlgo
+
+var packageNameTemplate string = `package main\n`
+
+var mainImportTemplate string = `import (
+	"encoding/json"
+	"fmt"
+	_ "github.com/go-sql-driver/mysql"
+	"github.com/graphql-go/graphql"
+	"github.com/jinzhu/gorm"
+	_ "github.com/jinzhu/gorm/dialects/mysql"
+	"log"
+	"net/http"
+)\n`
